ai/requirements.txt
opencv-python
torch
ultralytics
Pillow
numpy